[add]: cart

+++ src/org/experis/shop/Smartphone.java
@@ -4,10 +4,10 @@ import java.math.BigDecimal;
 
 public class Smartphone extends Product{
 
-    private int codeIMEI;
+    private long codeIMEI;
     private int memory;
 
-    public Smartphone(String name, String brand, BigDecimal price, BigDecimal vat, int codeIMEI, int memory) {
+    public Smartphone(String name, String brand, BigDecimal price, BigDecimal vat, long codeIMEI, int memory) {
         super(name, brand, price, vat);
         this.codeIMEI = codeIMEI;
         this.memory = memory;
